beer: 🍻 move to `./themes/`, no need for `./themes/default` yet
Also very sober just more testing of conventional-commits from gitmoji
