+++ lib/createState.js
@@ -30,7 +30,7 @@ const createState = (config = {}) => {
   // @todo _switch_ this up
   if (state.config.theme && state.config.theme === 'gitmoji') {
     // eslint-disable-next-line global-require
-    const {list, types} = require('./getGitmoji');
+    const {list, types} = require('./themes/gitmoji');
     state.config.breakingChangePrefix = '💥️ ';
     state.config.list = list;
     state.config.types = types;
