+++ lib/themes/gitmoji.js
@@ -1,7 +1,7 @@
 /* eslint-disable object-property-newline */
 // eslint-disable-next-line id-match
 const _find = require('lodash/find');
-const gitmojis = require('../data/gitmojis.json').gitmojis;
+const items = require('../../data/gitmojis.json').gitmojis;
 
 // @note pseudo-map to conventional-commits
 const rewrites = [
@@ -87,27 +87,27 @@ const types = {
   },
 };
 
-const getGitmoji = async () => {
+const gitmoji = async () => {
   // eslint-disable-next-line array-callback-return
-  await gitmojis.map((gitmoji) => {
-    const rewrite = _find(rewrites, {from: gitmoji.name});
+  await items.map((item) => {
+    const rewrite = _find(rewrites, {from: item.name});
     list.push(rewrite.to);
     types[rewrite.to] = {
-      code: gitmoji.code,
-      description: gitmoji.description,
-      emoji: gitmoji.emoji,
-      entity: gitmoji.entity,
+      code: item.code,
+      description: item.description,
+      emoji: item.emoji,
+      entity: item.entity,
       hidden: false,
-      name: gitmoji.name,
-      release: gitmoji.semver || null,
-      section: gitmoji.description,
-      semver: gitmoji.semver,
+      name: item.name,
+      release: item.semver || null,
+      section: item.description,
+      semver: item.semver,
       value: rewrite.to,
     };
   });
 };
 
-getGitmoji();
+gitmoji();
 
 module.exports = {
   list,
